==> solubility/3-NutrientSolubility-pH-corrected.Rmd <==
---
title: "Nutrient Solubility"
subtitle: "Theoretical values corrected by pH"
author: "Anıl A. Tellbüscher"
date: '2022-07-06'
output: html_document 
---

```{r include=FALSE}
library(XLConnect)
library(tidyverse)
```



# General
While comparing nutrient concentrations in aquaponic systems reported in literature with theoretical solubilities based on calculations using the solubility product constant $K_{sp}$, reported concentrations are approximately three orders of magnitude higher in case of Ca, Mg, Fe, Mn, Cu, Zn, and P. This could be due to the pH-dependence of the solubility of some salts.

For instance, orthophosphates undergo a number of acid-base reactions when introduced into aquatic systems. The chemical equilibria are depending on the system pH. In hydroponic systems, the recommended pH ranges around 6.0, while it is recommended to maintain a pH above 7.0 in aquaculture systems, including the aquaculture unit of on-demand coupled aquaponic systems. Considering the pH when calculating solubilities might thus lead to better, more precise results, because the pH is calculated on a logarithmic scale, which means that a change in the pH by one unit results in a change of the hydronium ion concentration by one order of magnitude. 


# Preparation of data analysis

## 1.1 Load functions for species calculation

Some functions were written to calculate the true concentration of the precipitation-causing species in dependence of their total concentration in the system and the system pH. The acidity constants $K_{a}$ were taken from literature.

```{r Load functions, include=FALSE}
# Phosphate speciation
source('R/speciation - c_T-PO4.R')

# Carbonate speciation
source('R/speciation - pCO2-CO3.R')

# Hydroxide speciation
source('R/speciation - OH.R')

# Sulfate speciation
source('R/speciation - c_T-SO4.R')
```





## 1.2 Load data

Rawdata is gathered by skimming relevant literature for data that can be used for a meta-analysis.

```{r Load workbook, include=FALSE}
# Load workbook
masterfile_nutrients <- loadWorkbook(filename = "data/Aquaponics - Masterfile Nutrients.xlsx")
```



```{r Read worksheets, include=FALSE}
# Read worksheet - General information
general_info <- readWorksheet(masterfile_nutrients, sheet = 'general')

# Read worksheet - Nutrient concentrations in pre-filtered observations
empirical_data <- readWorksheet(masterfile_nutrients, sheet = 'waterAVERAGE')

# Load preprocessed data
solubility_data <- read.csv('data/solubilityData_for3-NutrientSolubilityCorrected.csv')

#print(empirical_data)-
```





```{r Process solubility data, include=FALSE}
solubility_data <- solubility_data %>% 
  select(-c(X, name, cation, anion, densH2O, massPercent)) %>% 
  filter(!is.na(Ks0) & chemAnion != "S" & chemAnion != "SO3") %>% 
  filter(chemFormula != "Fe(OH)2" & chemFormula != "Fe(OH)3" & chemFormula != "FeCO3")

solubility_data

# Write solubility data
write.csv(solubility_data, 'data/solubilityData_for4-NutrientComparison.csv')
```


For correction of the solubilities by pH, the total concentrations $c_{T}$ of anions such as phosphorus in form of phosphates and sulfur in form of sulfates reported in the publications are recalculated based on the reported pH. The concentration of hydroxide is directly derived from the pH. Furthermore, the concentration of carbonate at the given pH is calculated, based on the assumption that the water is saturated with carbon dioxide. This yields the concentration of the dissolved species that is causing precipitation. Given this value and the solubility constants, the possible free concentrations of the cations that are forming precipitates with the mentioned anions is calculated.




# 2. pH-correction - 1st iteration

All concentrations are initially reported in mg/L.

```{r Create procdata and recalculate species, echo=FALSE}
procdata <- empirical_data %>% 
  
  # Rename columns
  rename(
    'NH4-N' = 'N.NH4_mgL', 
    'NO2-N' = 'N.NO2_mgL', 
    'NO3-N' = 'N.NO3_mgL', 
    'PO4-P' = 'P.PO4_mgL', 
    'K' = 'K_mgL', 
    'Ca' = 'Ca_mgL', 
    'Mg' = 'Mg_mgL', 
    'SO4-S' = 'S.SO4_mgL', 
    'B' = 'B_mgL', 
    'Fe' = 'Fe_mgL', 
    'Mn' = 'Mn_mgL', 
    'Cu' = 'Cu_mgL', 
    'Zn' = 'Zn_mgL', 
    'Na' = 'Na_mgL', 
    'Mo' = 'Mo_mgL', 
    'Ni' = 'Ni_mgL'
    ) %>% 

  select(-ends_with(c('const', 'belowLimit'))) %>% 

  # Replace NAs by 0
  mutate_at(vars(`NH4-N`, `NO2-N`, `NO3-N`), tidyr::replace_na, 0) %>%
  
  mutate(
    # Calculate Total Inorganic Nitrogen by summing NH4, NO2 and NO3
    TIN = `NH4-N` + `NO2-N` + `NO3-N`,
    # Calculate [PO4]
    species_PO4_molL = calc_PO4(c_tot = `PO4-P`, pH = pH, form_P = "PO4-P", unit = "mg"),
    # Calculate [CO3]
    species_CO3_molL = calc_CO3(pH = pH),
    # Calculate [OH]
    species_OH_molL = calc_OH(pH = pH),
    # Calculate [SO4]
    species_SO4_molL = calc_SO4(c_tot = `SO4-S`, pH = pH, form_S = "SO4-S", unit = "mg")
    ) %>% 
  print()
```






```{r Convert procdata to mol}
# Import molar weight data
molarWeight <- read.csv2(file = 'data/molarWeight.csv')

# 
procdata <- procdata %>% 

  rename('S' = 'SO4-S', 'N' = 'TIN') %>% 
  
  pivot_longer( # Convert into longtable
    cols = K:N,
    names_to = 'nutrient',
    values_to = 'concentration_mgL'
    ) %>% 
  left_join(molarWeight, by = c('nutrient' = 'element')) %>% 
  
  mutate(
    g_per_mol = as.numeric(g_per_mol), # Convert from character into numeric
    concentration_molL = concentration_mgL/(1e3 * g_per_mol) # Convert from mg/L into mol/L
    ) %>% 
  
  select(-c(concentration_mgL, g_per_mol)) %>% # Remove unnecessary columns
  
  pivot_wider( # Convert into widetable
    names_from = 'nutrient',
    values_from = 'concentration_molL'
    ) %>% 
  
  print()
```





```{r Recalculate solubilities}
procdata <- procdata %>% 
  mutate(
    # CaCO3
    Ca_in_CaCO3 = solubility_data$Ks0[which(solubility_data$chemFormula == "CaCO3",)] / species_CO3_molL,
    Ca_found_greater_CaCO3 = Ca > Ca_in_CaCO3,
    CaCO3_diff = Ca - Ca_in_CaCO3,
    CaCO3_prop = Ca / Ca_in_CaCO3,
    
    # Ca(OH)2
    Ca_in_CaOH2 = solubility_data$Ks0[which(solubility_data$chemFormula == "Ca(OH)2",)] / species_OH_molL^2,
    Ca_found_greater_CaOH2 = Ca > Ca_in_CaOH2,
    CaOH2_diff = Ca - Ca_in_CaOH2,
    CaOH2_prop = Ca / Ca_in_CaOH2,
    
    # Ca3PO42
    Ca_in_Ca3PO42 = (solubility_data$Ks0[which(solubility_data$chemFormula == "Ca3(PO4)2",)] / species_PO4_molL^2)^(1/3),
    Ca_found_greater_Ca3PO42 = Ca > Ca_in_Ca3PO42,
    Ca3PO42_diff = Ca - Ca_in_Ca3PO42,
    Ca3PO42_prop = Ca / Ca_in_Ca3PO42,

    
    # CaSO4
    Ca_in_CaSO4 = solubility_data$Ks0[which(solubility_data$chemFormula == "CaSO4",)] / species_SO4_molL,
    Ca_found_greater_CaSO4 = Ca > Ca_in_CaSO4,
    CaSO4_diff = Ca - Ca_in_CaSO4,
    CaSO4_prop = Ca / Ca_in_CaSO4,
    
    # MgCO3
    Mg_in_MgCO3 = solubility_data$Ks0[which(solubility_data$chemFormula == "MgCO3",)] / species_CO3_molL,
    Mg_found_greater_MgCO3 = Mg > Mg_in_MgCO3,
    MgCO3_diff = Mg - Mg_in_MgCO3,
    MgCO3_prop = Mg / Mg_in_MgCO3,
    
    # Mg(OH)2
    Mg_in_MgOH2 = solubility_data$Ks0[which(solubility_data$chemFormula == "Mg(OH)2",)] / species_OH_molL^2,
    Mg_found_greater_MgOH2 = Mg > Mg_in_MgOH2,
    MgOH2_diff = Mg - Mg_in_MgOH2,
    MgOH2_prop = Mg / Mg_in_MgOH2,

    
    # Mg3PO42
    Mg_in_Mg3PO42 = (solubility_data$Ks0[which(solubility_data$chemFormula == "Mg3(PO4)2",)] / species_PO4_molL^2)^(1/3),
    Mg_found_greater_Mg3PO42 = Mg > Mg_in_Mg3PO42,
    Mg3PO42_diff = Mg - Mg_in_Mg3PO42,
    Mg3PO42_prop = Mg / Mg_in_Mg3PO42,

    # FePO4
    Fe_in_FePO4 = solubility_data$Ks0[which(solubility_data$chemFormula == "FePO4 x 2 H2O",)] / species_PO4_molL,
    Fe_found_greater_FePO4 = Fe > Fe_in_FePO4,
    FePO4_diff = Fe - Fe_in_FePO4,
    FePO4_prop = Fe / Fe_in_FePO4,
    
    # MnCO3
    Mn_in_MnCO3 = solubility_data$Ks0[which(solubility_data$chemFormula == "MnCO3",)] / species_CO3_molL,
    Mn_found_greater_MnCO3 = Mn > Mn_in_MnCO3,
    MnCO3_diff = Mn - Mn_in_MnCO3,
    MnCO3_prop = Mn / Mn_in_MnCO3,
    
    # Cu3PO42
    Cu_in_Cu3PO42 = (solubility_data$Ks0[which(solubility_data$chemFormula == "Cu3(PO4)2",)] / species_PO4_molL^2)^(1/3),
    Cu_found_greater_Cu3PO42 = Cu > Cu_in_Cu3PO42,
    Cu3PO42_diff = Cu - Cu_in_Cu3PO42,
    Cu3PO42_prop = Cu / Cu_in_Cu3PO42,
    
    # ZnCO3
    Zn_in_ZnCO3 = solubility_data$Ks0[which(solubility_data$chemFormula == "ZnCO3",)] / species_CO3_molL,
    Zn_found_greater_ZnCO3 = Zn > Zn_in_ZnCO3,
    ZnCO3_diff = Zn - Zn_in_ZnCO3,
    ZnCO3_prop = Zn / Zn_in_ZnCO3,
    
    # Zn(OH)2
    Zn_in_ZnOH2 = solubility_data$Ks0[which(solubility_data$chemFormula == "Zn(OH)2",)] / species_OH_molL^2,
    Zn_found_greater_ZnOH2 = Zn > Zn_in_ZnOH2,
    ZnOH2_diff = Zn - Zn_in_ZnOH2,
    ZnOH2_prop = Zn / Zn_in_ZnOH2,
    
    # NiCO3
    Ni_in_NiCO3 = solubility_data$Ks0[which(solubility_data$chemFormula == "NiCO3",)] / species_CO3_molL,
    Ni_found_greater_NiCO3 = Ni > Ni_in_NiCO3,
    NiCO3_diff = Ni - Ni_in_NiCO3,
    NiCO3_prop = Ni / Ni_in_NiCO3,
    
    # Ni(OH)2
    Ni_in_NiOH2 = solubility_data$Ks0[which(solubility_data$chemFormula == "Ni(OH)2",)] / species_OH_molL^2,
    Ni_found_greater_NiOH2 = Ni > Ni_in_NiOH2,
    NiOH2_diff = Ni - Ni_in_NiOH2,
    NiOH2_prop = Ni / Ni_in_NiOH2,
    
    # Ni3PO42
    Ni_in_Ni3PO42 = (solubility_data$Ks0[which(solubility_data$chemFormula == "Ni3(PO4)2",)] / species_PO4_molL^2)^(1/3),
    Ni_found_greater_Ni3PO42 = Ni > Ni_in_Ni3PO42,
    Ni3PO42_diff = Ni - Ni_in_Ni3PO42,
    Ni3PO42_prop = Ni / Ni_in_Ni3PO42
  ) %>% 
  
  write_csv(file = 'data/interm/nutrientSolubility_pHcorrected.csv') %>% 
  
  print()
```  

The following compounds were not considered:

\begin{itemize}
  \item Fe(OH)3: Labile
  \item Fe2(CO3)3: Labile; rapid decomposition reaction to Fe2O3 + 3 CO2
\end{itemize}


```{r}
procdata %>% 
  select(Ca_in_CaCO3:Ni_in_Ni3PO42) %>% 
  print()
```


**Are empirical values above the calculated solubilities?**

|   |CO3|OH |PO4|SO4|
|---|---|---|---|---|
|Ca |yes|no |yes|no |
|Mg |no |no |no |-  |
|Fe |-  |-  |yes|-  |
|Mn |yes|-  |-  |-  |
|Cu |-  |-  |yes|-  |
|Zn |yes|no |-  |-  |
|Ni |-  |-  |-  |-  |




## 2.3. Plot data

### 2.3.1 Create expectation_data

```{r}
#procdata %>% 
  
```


### 2.3.2 Create plotdata

```{r}
plotdata <- procdata %>% 
  select(ends_with('_prop')) %>% 
  pivot_longer(
    cols = everything(),
    names_to = 'substance',
    values_to = 'proportion'
  ) %>% 
  mutate(
    substance = sub('_prop', '', substance)
  ) %>% 
  filter(
    proportion != Inf
    ) %>% 
  print()
```



### 2.3.3 Create plot




```{r}
plotdata %>% 
  #filter(proportion < 2) %>% 
  mutate(
    substance = as.factor(substance),
    treshold_label = if_else(proportion > 1, true = '2', false = '1')
    ) %>% 
  ggplot(aes(
    x = substance,
    y = proportion,
    fill = treshold_label
    )) + 
    geom_boxplot() + 
    geom_hline(yintercept = 1, linetype = 'dashed', color = 'red', alpha = 0.75) +
    scale_y_log10() + 
    #facet_wrap(facets = vars(substance), scales = 'free') + 
    labs(
      x = 'Empirical solubility compared with solubility of salt',
      y = 'log(proportion)'
    ) + 
    theme_minimal() + 
    theme(
      axis.text.x = element_text(angle = 45),
      legend.position = 'none'
      )
```



The results can be explained as follows:

*Carbonates:*
It appears as the concentration of calcium is much higher than possible if carbonates were limiting the concentration in the system. This is true for CaCO3, MnCO3 and ZnCO3. However, it was assumed that the water is in equilibrium with atmospheric CO2, which is oftentimes not the case in aquaculture systems. Instead, the water is degassed. Currently, the assumed concentration of CO2 in water is [CO2(aq)] $\approx$ 70 mg/L. According to Terjesen et al. (2013), a more realistic value would be 5 mg/L. Thus, in the second iteration, the CO2 concentration will be reduced.

*Phosphates:*
The concentration of Ca, Fe and Cu is higher than theoretically possible if precipitation of phosphates was the concentration limiting factor. With respect to **Fe** and **Cu**, the reason for the comparably higher values could be the formation of complex molecules. Complexation can lead to a concentration increase of the affected metals by approximately three orders of magnitude, depending on the stability of the resulting complex. With respect to **Ca**, the reason for the elevated concentrations might be interactions with humic matter.





# 3. pH-correction - 2nd iteration

As mentioned, the concentration of CO2 will be reduced in this iteration by reducing the assumed partial pressure of CO2 from 0.054 to 0.0027 atm. This results in a decrease of the CO2 concentrtion in water by factor 20. All concentrations are initially reported in mg/L.

```{r Create procdata and recalculate species 2}
procdata <- empirical_data %>% 
  
  # Rename columns
  rename('NH4-N' = 'N.NH4_mgL', 'NO2-N' = 'N.NO2_mgL', 'NO3-N' = 'N.NO3_mgL', 'PO4-P' = 'P.PO4_mgL', 'K' = 'K_mgL', 'Ca' = 'Ca_mgL', 'Mg' = 'Mg_mgL', 'SO4-S' = 'S.SO4_mgL', 'B' = 'B_mgL', 'Fe' = 'Fe_mgL', 'Mn' = 'Mn_mgL', 'Cu' = 'Cu_mgL', 'Zn' = 'Zn_mgL', 'Na' = 'Na_mgL', 'Mo' = 'Mo_mgL', 'Ni' = 'Ni_mgL') %>% 

  select(-ends_with(c('const', 'belowLimit'))) %>% 

  # Replace NAs by 0
  mutate_at(vars(`NH4-N`, `NO2-N`, `NO3-N`), tidyr::replace_na, 0) %>%
  
  mutate(
    # Calculate Total Inorganic Nitrogen by summing NH4, NO2 and NO3
    TIN = `NH4-N` + `NO2-N` + `NO3-N`,
    # Calculate [PO4]
    species_PO4_molL = calc_PO4(c_tot = `PO4-P`, pH = pH, form_P = "PO4-P", unit = "mg"),
    # Calculate [CO3]
    species_CO3_molL = calc_CO3(pH = pH, pCO2 = 0.0027),
    # Calculate [OH]
    species_OH_molL = calc_OH(pH = pH),
    # Calculate [SO4]
    species_SO4_molL = calc_SO4(c_tot = `SO4-S`, pH = pH, form_S = "SO4-S", unit = "mg")
    ) %>% 
  print()
```





```{r Convert procdata to mol 2}
# Read molar weight data
molarWeight <- read.csv2(file = 'data/molarWeight.csv')

# 
procdata <- procdata %>% 

  rename('S' = 'SO4-S', 'N' = 'TIN') %>% 
  
  pivot_longer( # Convert into longtable
    cols = K:N,
    names_to = 'nutrient',
    values_to = 'concentration_mgL'
    ) %>% 
  left_join(molarWeight, by = c('nutrient' = 'element')) %>% 
  
  mutate(
    g_per_mol = as.numeric(g_per_mol), # Convert from character into numeric
    concentration_molL = concentration_mgL/(1e3 * g_per_mol) # Convert from mg/L into mol/L
    ) %>% 
  
  select(-c(concentration_mgL, g_per_mol)) %>% # Remove unnecessary columns
  
  pivot_wider( # Convert into widetable
    names_from = 'nutrient',
    values_from = 'concentration_molL'
    ) %>% 
  
  print()
```





```{r Recalculate solubilities 2}
procdata <- procdata %>% 
  mutate(
    # CaCO3
    Ca_in_CaCO3 = solubility_data$Ks0[which(solubility_data$chemFormula == "CaCO3",)] / species_CO3_molL,
    Ca_found_greater_CaCO3 = Ca > Ca_in_CaCO3,
    CaCO3_diff = Ca - Ca_in_CaCO3,
    CaCO3_prop = Ca / Ca_in_CaCO3,
    
    # Ca(OH)2
    Ca_in_CaOH2 = solubility_data$Ks0[which(solubility_data$chemFormula == "Ca(OH)2",)] / species_OH_molL^2,
    Ca_found_greater_CaOH2 = Ca > Ca_in_CaOH2,
    CaOH2_diff = Ca - Ca_in_CaOH2,
    CaOH2_prop = Ca / Ca_in_CaOH2,
    
    # Ca3PO42
    Ca_in_Ca3PO42 = (solubility_data$Ks0[which(solubility_data$chemFormula == "Ca3(PO4)2",)] / species_PO4_molL^2)^(1/3),
    Ca_found_greater_Ca3PO42 = Ca > Ca_in_Ca3PO42,
    Ca3PO42_diff = Ca - Ca_in_Ca3PO42,
    Ca3PO42_prop = Ca / Ca_in_Ca3PO42,

    
    # CaSO4
    Ca_in_CaSO4 = solubility_data$Ks0[which(solubility_data$chemFormula == "CaSO4",)] / species_SO4_molL,
    Ca_found_greater_CaSO4 = Ca > Ca_in_CaSO4,
    CaSO4_diff = Ca - Ca_in_CaSO4,
    CaSO4_prop = Ca / Ca_in_CaSO4,
    
    # MgCO3
    Mg_in_MgCO3 = solubility_data$Ks0[which(solubility_data$chemFormula == "MgCO3",)] / species_CO3_molL,
    Mg_found_greater_MgCO3 = Mg > Mg_in_MgCO3,
    MgCO3_diff = Mg - Mg_in_MgCO3,
    MgCO3_prop = Mg / Mg_in_MgCO3,
    
    # Mg(OH)2
    Mg_in_MgOH2 = solubility_data$Ks0[which(solubility_data$chemFormula == "Mg(OH)2",)] / species_OH_molL^2,
    Mg_found_greater_MgOH2 = Mg > Mg_in_MgOH2,
    MgOH2_diff = Mg - Mg_in_MgOH2,
    MgOH2_prop = Mg / Mg_in_MgOH2,

    
    # Mg3PO42
    Mg_in_Mg3PO42 = (solubility_data$Ks0[which(solubility_data$chemFormula == "Mg3(PO4)2",)] / species_PO4_molL^2)^(1/3),
    Mg_found_greater_Mg3PO42 = Mg > Mg_in_Mg3PO42,
    Mg3PO42_diff = Mg - Mg_in_Mg3PO42,
    Mg3PO42_prop = Mg / Mg_in_Mg3PO42,

    # FePO4
    Fe_in_FePO4 = solubility_data$Ks0[which(solubility_data$chemFormula == "FePO4 x 2 H2O",)] / species_PO4_molL,
    Fe_found_greater_FePO4 = Fe > Fe_in_FePO4,
    FePO4_diff = Fe - Fe_in_FePO4,
    FePO4_prop = Fe / Fe_in_FePO4,
    
    # MnCO3
    Mn_in_MnCO3 = solubility_data$Ks0[which(solubility_data$chemFormula == "MnCO3",)] / species_CO3_molL,
    Mn_found_greater_MnCO3 = Mn > Mn_in_MnCO3,
    MnCO3_diff = Mn - Mn_in_MnCO3,
    MnCO3_prop = Mn / Mn_in_MnCO3,
    
    # Cu3PO42
    Cu_in_Cu3PO42 = (solubility_data$Ks0[which(solubility_data$chemFormula == "Cu3(PO4)2",)] / species_PO4_molL^2)^(1/3),
    Cu_found_greater_Cu3PO42 = Cu > Cu_in_Cu3PO42,
    Cu3PO42_diff = Cu - Cu_in_Cu3PO42,
    Cu3PO42_prop = Cu / Cu_in_Cu3PO42,
    
    # ZnCO3
    Zn_in_ZnCO3 = solubility_data$Ks0[which(solubility_data$chemFormula == "ZnCO3",)] / species_CO3_molL,
    Zn_found_greater_ZnCO3 = Zn > Zn_in_ZnCO3,
    ZnCO3_diff = Zn - Zn_in_ZnCO3,
    ZnCO3_prop = Zn / Zn_in_ZnCO3,
    
    # Zn(OH)2
    Zn_in_ZnOH2 = solubility_data$Ks0[which(solubility_data$chemFormula == "Zn(OH)2",)] / species_OH_molL^2,
    Zn_found_greater_ZnOH2 = Zn > Zn_in_ZnOH2,
    ZnOH2_diff = Zn - Zn_in_ZnOH2,
    ZnOH2_prop = Zn / Zn_in_ZnOH2,
    
    # NiCO3
    Ni_in_NiCO3 = solubility_data$Ks0[which(solubility_data$chemFormula == "NiCO3",)] / species_CO3_molL,
    Ni_found_greater_NiCO3 = Ni > Ni_in_NiCO3,
    NiCO3_diff = Ni - Ni_in_NiCO3,
    NiCO3_prop = Ni / Ni_in_NiCO3,
    
    # Ni(OH)2
    Ni_in_NiOH2 = solubility_data$Ks0[which(solubility_data$chemFormula == "Ni(OH)2",)] / species_OH_molL^2,
    Ni_found_greater_NiOH2 = Ni > Ni_in_NiOH2,
    NiOH2_diff = Ni - Ni_in_NiOH2,
    NiOH2_prop = Ni / Ni_in_NiOH2,
    
    # Ni3PO42
    Ni_in_Ni3PO42 = (solubility_data$Ks0[which(solubility_data$chemFormula == "Ni3(PO4)2",)] / species_PO4_molL^2)^(1/3),
    Ni_found_greater_Ni3PO42 = Ni > Ni_in_Ni3PO42,
    Ni3PO42_diff = Ni - Ni_in_Ni3PO42,
    Ni3PO42_prop = Ni / Ni_in_Ni3PO42
  ) %>% 
  print()
```  



The following compounds were not considered:

\begin{itemize}
  \item Fe(OH)3: Labile
  \item Fe2(CO3)3: Labile; rapid decomposition reaction to Fe2O3 + 3 CO2
\end{itemize}



```{r}
procdata %>% 
  select(Ca_in_CaCO3:Ni_in_Ni3PO42) %>% 
  print()
```


**Are empirical values above the calculated solubilities?**

|   |CO3|OH |PO4|SO4|
|---|---|---|---|---|
|Ca |yes|no |yes|no |
|Mg |no |no |no |-  |
|Fe |-  |-  |yes|-  |
|Mn |yes|-  |-  |-  |
|Cu |-  |-  |yes|-  |
|Zn |yes|no |-  |-  |
|Ni |-  |-  |-  |-  |




## 3.3 Plot data

### 3.3.1 Create expectation_data

```{r}
#procdata %>% 
  
```


### 3.3.2 Create plotdata

```{r}
plotdata <- procdata %>% 
  select(ends_with('_prop')) %>% 
  pivot_longer(
    cols = everything(),
    names_to = 'substance',
    values_to = 'proportion'
  ) %>% 
  mutate(
    substance = sub('_prop', '', substance)
  ) %>% 
  filter(
    proportion != Inf
    ) %>% 
  print()
```



### 3.3.3 Create plot




```{r}
plotdata %>% 
  #filter(proportion < 2) %>% 
  mutate(
    substance = as.factor(substance),
    treshold_label = if_else(proportion > 1, true = '2', false = '1')
    ) %>% 
  ggplot(aes(
    x = substance,
    y = proportion,
    fill = treshold_label
    )) + 
    geom_boxplot() + 
    geom_hline(yintercept = 1, linetype = 'dashed', color = 'red', alpha = 0.75) +
    scale_y_log10() + 
    #facet_wrap(facets = vars(substance), scales = 'free') + 
    labs(
      x = 'Empirical solubility compared with solubility of salt',
      y = 'log(proportion)'
    ) + 
    theme_minimal() + 
    theme(
      axis.text.x = element_text(angle = 45),
      legend.position = 'none'
      )
```



Now life is much better!
*Carbonates:*
Even though the concentrations for CaCO3 and ZnCO3 are still above the theoretical values, this can again be a matter of low CO2 concentrations in the systems, for instance as a result of low stocking densities. The value of 5 mg/L of CO2 was reported for a system with a stocking density of 25 kg/m3. Considering that supernatants from aerobic digesters are part of the dataset as well, it might still be the case that the carbonate concentration is overestimated.
